Use state machine based simulation by default
